feature(get a single post): user should get a single post
