+++ Src/App/Controllers/blogControllers.js
@@ -1,13 +1,12 @@
 /* eslint-disable linebreak-style */
 import ResponseHandler from '../helpers/responseHandler';
-import JwtAuth from '../helpers/jwtHelper';
-// import BlogController from './userController';
+import ExtractToken from '../helpers/extractToken';
 import BlogServices from '../../database/services/blogService';
 
 /** @class BlogControllers */
 class BlogControllers {
   /**
-   * signup user in the database
+   * createBlog create a new blog and store it in the database
    * @static
    * @param {*} req
    * @param {*} res
@@ -17,17 +16,9 @@ class BlogControllers {
    */
   static async createBlog(req, res, next) {
     try {
-      const token = req.headers.authorization;
-      if (!token) {
-        ResponseHandler.error(res, 404, 'Access Denied!');
-        return next();
-      }
-      const extractToken = token.split(' ')[1];
-      const userId = await JwtAuth.getCurrentUserId(extractToken);
-      if (!userId) {
-        ResponseHandler.error(res, 404, 'Not Authorized!');
-        return next();
-      }
+      // get userId fom the token
+      const userId = await ExtractToken.extractAccessToken(req, res, next);
+
       const { title, description } = req.body;
 
       const blogData = {
@@ -65,20 +56,44 @@ class BlogControllers {
   static async getAllBlog(req, res, next) {
     try {
       // get the userId from the token
-      const token = req.headers.authorization;
-      if (!token) {
-        ResponseHandler.error(res, 404, 'Access Denied!');
-        return next();
-      }
-
-      const extractToken = token.split(' ')[1];
-      const userId = await JwtAuth.getCurrentUserId(extractToken);
-      if (!userId) {
-        ResponseHandler.error(res, 404, 'Not Authorized!');
-        return next();
-      }
+      const userId = await ExtractToken.extractAccessToken(req, res, next);
 
       const result = await BlogServices.getAllP(userId);
+
+      if (!result) {
+        ResponseHandler.error(res, 400, 'Not found!');
+        return next();
+      }
+      // if there is no any blog registered with this userId
+      if (result === null) {
+        ResponseHandler.error(res, 400, 'You have no blog yet');
+      } else {
+        // return the response to the client
+        ResponseHandler.success(res, 200, '', { result });
+      }
+    } catch (error) {
+      ResponseHandler.error(res, 400, error.message);
+    }
+  }
+
+  /**
+   * getSingleBlog in the database by userId
+   * @static
+   * @param {*} req
+   * @param {*} res
+   * @param {*} next
+   * @returns {object} data
+   * @memberof BlogController
+   */
+  static async getSingleBlog(req, res, next) {
+    try {
+      // get the userId from the token
+      // const userId = await ExtractToken.extractAccessToken(req, res, next);
+
+      // Get a blog Id from the parameters
+      const blogId = req.params.post_id;
+
+      const result = await BlogServices.getSingleB(blogId);
 
       if (!result) {
         ResponseHandler.error(res, 400, 'Not found!');
