+++ server.js
@@ -3,7 +3,7 @@
 import 'dotenv/config';
 import app from './src/app/app';
 
-const PORT = process.env.PORT || 3020;
+const PORT = process.env.PORT || 3022;
 
 app.listen(PORT, () => {
   process.stdout.write(`server listening on: http://localhost:${PORT}`);
